Major refactor to accomodate cryptonote::core
- We don't want to create a copy, like our upstream did with the daemon vm_hack, so we access core similarly to the rpc implementation.  I'm sure this isn't completely correct, but its a start for getting the core stuff in such a way that we can access the db, without causing problems (hopefully)
- Within m_core we can also access checkpointing code, so this should accomplish that as well.
- This should at least give us all of the tools we need to monitor blockchain, create and modify a komodo_state file, and we can use blur-notary-server-rpc to write the notarization transactions
